requirements.txt
torch>=1.9.0
transformers>=4.5.0
pandas>=1.2.0
numpy>=1.19.5
tqdm>=4.62.0
PyYAML>=5.4.1